+++ lib/trailblazer/operation/pundit.rb
@@ -8,13 +8,13 @@ class Trailblazer::Operation
       end
 
       def self.build(*args, &block)
-        Permission.new(*args, &block)
+        Condition.new(*args, &block)
       end
 
       # This can be subclassed for other policy strategies, e.g. non-pundit Authsome.
       # NOTE: using a class here is faster than a simple proc: https://twitter.com/apotonick/status/791162989692891136
       #
-      class Permission
+      class Condition
         def initialize(policy_class, action)
           @policy_class, @action = policy_class, action
         end