+++ test/docs/policy_test.rb
@@ -1,0 +1,2 @@
+# add anything to result
+# insert any options into constructor